LayoutTests/imported/w3c/web-platform-tests/navigation-api/scroll-behavior/manual-scroll-before-after-transition-commit-expected.txt


FAIL scroll: scroll() before commit() assert_equals: expected 0 but got 431
